requirements.txt
curl_cffi
beautifulsoup4
python-dotenv
pandas
matplotlib
requests
pytest
